fix: integer value serialized as double in export
Cast from `long` to `int` when writing the crate.
Otherwise the overloaded method called is with `double` rather than
`int`.

+++ src/main/java/ch/psi/ord/core/RoCrateExporter.java
@@ -125,8 +125,8 @@ public class RoCrateExporter {
               publicationBuilder.addProperty(formatScicatId("relatedPublications"), p);
             });
     publicationBuilder
-        .addProperty(formatScicatId("numberOfFiles"), publication.getNumberOfFiles())
-        .addProperty(formatScicatId("sizeOfArchive"), publication.getSizeOfArchive())
+        .addProperty(formatScicatId("numberOfFiles"), (int) publication.getNumberOfFiles())
+        .addProperty(formatScicatId("sizeOfArchive"), (int) publication.getSizeOfArchive())
         .addProperty(formatScicatId("scicatUser"), publication.getScicatUser());
 
     DataEntity publicationEntity = publicationBuilder.build();
